Enable parallel tests
