+++ source/Halibut.Tests/Properties/AssemblyInfo.cs
@@ -1,7 +1,10 @@
 using System;
 using System.Reflection;
+using NUnit.Framework;
 
 // Information about this assembly is defined by the following attributes. 
 // Change them to the values specific to your project.
 
 [assembly: AssemblyTitle("Halibut.Tests")]
+[assembly: Parallelizable(ParallelScope.All)]
+[assembly: FixtureLifeCycle(LifeCycle.InstancePerTestCase)]
