+++ source/Halibut.Tests/TestServices/EchoService.cs
@@ -16,11 +16,8 @@ namespace Halibut.Tests.TestServices
             throw new DivideByZeroException();
         }
 
-        public static Action OnLongRunningOperation { get; set; }
-
         public int LongRunningOperation()
         {
-            OnLongRunningOperation();
             Thread.Sleep(10000);
             return 12;
         }
